% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\name{AsLazyMatrix}
\alias{AsLazyMatrix}
\title{Convert matrix or Matrix or similar to LazyMatrix}
\usage{
AsLazyMatrix(X)
}
\arguments{
\item{X}{regular matrix}
}
\description{
Convert matrix or Matrix or similar to LazyMatrix
}
