% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{methods}
\name{colMeans,LazyMatrix-method}
\alias{colMeans,LazyMatrix-method}
\title{Get means of columns for LazyMatrix.}
\usage{
\S4method{colMeans}{LazyMatrix}(x)
}
\arguments{
\item{x}{LazyMatrix}
}
\description{
Get means of columns for LazyMatrix.
}
