% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{methods}
\name{colSums,LazyMatrix-method}
\alias{colSums,LazyMatrix-method}
\title{Get sums of columns for LazyMatrix.}
\usage{
\S4method{colSums}{LazyMatrix}(x)
}
\arguments{
\item{x}{LazyMatrix}
}
\description{
Get sums of columns for LazyMatrix.
}
