% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\name{dim,LazyMatrix-method}
\alias{dim,LazyMatrix-method}
\title{What size is my matrix?}
\usage{
\S4method{dim}{LazyMatrix}(x)
}
\arguments{
\item{x}{LazyMatrix}
}
\description{
What size is my matrix?
}
