% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval.R, R/MatrixLazyEval_core.R
\name{EvaluateLazyMatrix}
\alias{EvaluateLazyMatrix}
\alias{EvaluateLazyMatrix}
\title{Compute a matrix product.}
\usage{
EvaluateLazyMatrix(M, RIGHT = diag(1, nrow = ncol(M), ncol = ncol(M)),
  LEFT = diag(1, nrow = nrow(M), ncol = nrow(M)))

EvaluateLazyMatrix(M, RIGHT = diag(1, nrow = ncol(M), ncol = ncol(M)),
  LEFT = diag(1, nrow = nrow(M), ncol = nrow(M)))
}
\arguments{
\item{M}{object of class MatrixLazyEval.}

\item{RIGHT}{y in xMy.}

\item{LEFT}{x in xMy.}

\item{M}{object of class LazyMatrix.}

\item{RIGHT}{y in xMy.}

\item{LEFT}{x in xMy.}
}
\description{
Compute a matrix product.

Compute a matrix product.
}
