% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_core.R
\name{ExtendLazyMatrix}
\alias{ExtendLazyMatrix}
\title{Left- and right-multiply a LazyMatrix lazily, without performing any calculations.}
\usage{
ExtendLazyMatrix(M, LEFT = NULL, RIGHT = NULL)
}
\arguments{
\item{M}{LazyMatrix to be multiplied}

\item{LEFT, RIGHT}{Lists of matrices to left- and right-multiply into M.
Everything is done left to right, so if these contain X, Y (LEFT) and Z, W (RIGHT), the result will represent
 XYMZW . If these are not lists, they will be wrapped in lists and named using deparse(substitute()).}
}
\description{
Left- and right-multiply a LazyMatrix lazily, without performing any calculations.
}
