% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_core.R
\name{ExtractElementsLazyMatrix}
\alias{ExtractElementsLazyMatrix}
\title{Evaluate elements of a LazyMatrix.}
\usage{
ExtractElementsLazyMatrix(M, i = 1:nrow(M), j = 1:ncol(M), drop = T)
}
\arguments{
\item{M}{LazyMatrix}

\item{i}{@param j Integer vectors.}
}
\description{
Evaluate elements of a LazyMatrix.
}
