% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\name{\%*\%,LazyMatrix,ANY-method}
\alias{\%*\%,LazyMatrix,ANY-method}
\title{}
\usage{
\S4method{\%*\%}{LazyMatrix,ANY}(x, y)
}
\arguments{
\item{x}{LazyMatrix}

\item{y}{This object must be able to interact with objects in the components slot of x via the usual matrix multiplication operator.

This function is EAGER: it triggers immediate calculations.}
}
\description{

}
