% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{methods}
\name{\%*\%,LazyMatrix,LazyMatrix-method}
\alias{\%*\%,LazyMatrix,LazyMatrix-method}
\title{Combine two LazyMatrices into another LazyMatrix.}
\usage{
\S4method{\%*\%}{LazyMatrix,LazyMatrix}(x, y)
}
\arguments{
\item{x, y}{LazyMatrix}
}
\description{
Combine two LazyMatrices into another LazyMatrix.
}
