% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval.R
\name{IsValidLazyMatrix}
\alias{IsValidLazyMatrix}
\title{MatrixLazyEval class}
\usage{
IsValidLazyMatrix(M)
}
\arguments{
\item{M}{object of class MatrixLazyEval.}
}
\description{
MatrixLazyEval class
}
\details{
eval_rule may only contain `LEFT`, `RIGHT`, names in `components`, and simple operations (`()`, `-*+`, `%*%`).
 `LEFT` and `RIGHT` specify the interface with the outside world: if your object is M and you perform M %*% x,
 then the value of x is substituted into RIGHT and LEFT is ignored.
 Like with R syntax, asterisk (`*`) means componentwise multiplication, and wrapped in percents
  (`%*%`) it means matrix multiplication.

Does it conform to the rules?
}
\section{Slots}{

\describe{
\item{\code{components}}{Named list containing matrices. Anything with a matrix multiplication operator ought to work.
Notably, you can put another instance of LazyMatrixEval.}

\item{\code{dim}}{Dimensions of the matrix represented here.}

\item{\code{eval_rule}}{Formula describing how to compute Mx or yM for this matrix (M) given x or y.}
}}

