% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_core.R
\name{IsValidRuleLazyMatrix}
\alias{IsValidRuleLazyMatrix}
\title{Does it conform to the rules?}
\usage{
IsValidRuleLazyMatrix(M)
}
\arguments{
\item{M}{object of class LazyMatrix.}
}
\description{
Does it conform to the rules?
}
