% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{class}
\name{LazyMatrix-class}
\alias{LazyMatrix-class}
\title{LazyMatrix class.}
\description{
LazyMatrix class.
}
\details{
eval_rule may only contain LEFT, RIGHT, names in \code{components}, and simple arithmetic
operations \code{(), -*+, \%*\% }.
 LEFT and RIGHT specify the interface with the outside world: if your object is M and you perform \code{M \%*\% x},
 then the value of x is substituted into RIGHT (and an identity matrix for LEFT).
 Like with typical R syntax, asterisk (*) means componentwise multiplication, and wrapped in percents
  (\code{\%*\%}) it means matrix multiplication.
}
\section{Slots}{

\describe{
\item{\code{components}}{Named list containing matrices. Anything with a matrix multiplication operator ought to work.
Notably, you can put another instance of LazyMatrixEval.}

\item{\code{dim}}{Dimensions of the matrix represented here.}

\item{\code{eval_rule}}{Length-1 character describing how to compute Mx or yM for this matrix (M) given x or y.}
}}

