% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\name{ncol,LazyMatrix-method}
\alias{ncol,LazyMatrix-method}
\title{How wide is my matrix?}
\usage{
\S4method{ncol}{LazyMatrix}(x)
}
\arguments{
\item{x}{LazyMatrix}
}
\description{
How wide is my matrix?
}
