% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{methods}
\name{nrow,LazyMatrix-method}
\alias{nrow,LazyMatrix-method}
\title{How tall is my matrix?}
\usage{
\S4method{nrow}{LazyMatrix}(x)
}
\arguments{
\item{x}{LazyMatrix}
}
\description{
How tall is my matrix?
}
