% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_pca.R
\name{PCALazyMatrix}
\alias{PCALazyMatrix}
\title{Compute randomized PCA.}
\usage{
PCALazyMatrix(M, ncomp, nproj_left = min(4 * ncomp, dim(M)[1]),
  nproj_right = min(4 * ncomp, dim(M)[2]))
}
\arguments{
\item{M}{object of class LazyMatrix.}

\item{ncomp}{number of components to compute}

\item{nproj_left}{dimension of random column space to project into}

\item{nproj_right}{dimension of random row    space to project into}
}
\description{
Compute randomized PCA.
}
