% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_convenience.R
\name{RankOneUpdateLazily}
\alias{RankOneUpdateLazily}
\title{Do a low-rank update to a matrix.}
\usage{
RankOneUpdateLazily(M, L, R)
}
\arguments{
\item{M}{Any matrix with a multiplication operator defined. Usually from the Matrix package.}

\item{L, R}{Matrices with dimensions such that $$M - LR$$ works.}
}
\description{
Do a low-rank update to a matrix.
}
