% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_convenience.R
\name{RegressOutLazily}
\alias{RegressOutLazily}
\title{Replace a matrix with residuals from an OLS regression.}
\usage{
RegressOutLazily(M, X)
}
\arguments{
\item{M}{Any matrix with a multiplication operator defined. Usually from the Matrix package.}

\item{X}{This must be some sort of typical matrix such that \code{solve(t(X) \%*\% X)} works.}
}
\description{
Replace a matrix with residuals from an OLS regression.
}
\details{
This function returns a LazyMatrix representing the residuals after regressing M on X.
If X is too big, though, you may not gain any efficiency.
}
