% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{methods}
\name{rowMeans,LazyMatrix-method}
\alias{rowMeans,LazyMatrix-method}
\title{Get means of rows for LazyMatrix.}
\usage{
\S4method{rowMeans}{LazyMatrix}(x)
}
\arguments{
\item{x}{LazyMatrix}
}
\description{
Get means of rows for LazyMatrix.
}
