% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{methods}
\name{rowSums,LazyMatrix-method}
\alias{rowSums,LazyMatrix-method}
\title{Get sums of rows for LazyMatrix.}
\usage{
\S4method{rowSums}{LazyMatrix}(x)
}
\arguments{
\item{x}{LazyMatrix}
}
\description{
Get sums of rows for LazyMatrix.
}
