% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_convenience.R
\name{ScaleLazily}
\alias{ScaleLazily}
\title{Scale rows or columns to a desired total.}
\usage{
ScaleLazily(M, do_rows = T, desired_sum = 1)
}
\arguments{
\item{M}{Any matrix with an \code{%*%} operator and nrow/ncol methods defined. Usually from the Matrix package.}

\item{do_rows}{If true, the rows are scaled so every row has the desired sum.
If false, the columns are centered so every column has the desired sum.}

\item{desired_sum}{Desired sum for each row or column.}
}
\description{
Scale rows or columns to a desired total.
}
