% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\name{[,LazyMatrix,ANY,ANY,ANY-method}
\alias{[,LazyMatrix,ANY,ANY,ANY-method}
\title{Abdicate responsibility for weird subsetting requests.}
\usage{
\S4method{[}{LazyMatrix,ANY,ANY,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{LazyMatrix}

\item{i, j, drop}{This failnice method gets dispatched if the methods above don't apply. 
It happens if the args are not integer or logical (also, integer disallowed for drop).}
}
\description{
Abdicate responsibility for weird subsetting requests.
}
