
    \docType{methods}
    \name{[,LazyMatrix,integer,missing,ANY-method}
    \alias{[,LazyMatrix,integer,missing,ANY-method}
    \title{Subset LazyMatrix}
    \usage{
    \S4method{[}{LazyMatrix,integer,missing,ANY}(x, i, j, drop = TRUE)
    }
    \arguments{
      \item{x}{LazyMatrix}
      \item{i, j, drop}{These args may be integer or logical (integer disallowed for drop).}
    }
    \description{
    "Subset as if from Matrix package."
    }
    