% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\name{summary,LazyMatrix-method}
\alias{summary,LazyMatrix-method}
\title{Show a few details on a LazyMatrix.}
\usage{
\S4method{summary}{LazyMatrix}(object)
}
\arguments{
\item{object}{LazyMatrix}
}
\description{
Show a few details on a LazyMatrix.
}
