% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{methods}
\name{t,LazyMatrix-method}
\alias{t,LazyMatrix-method}
\title{Transpose a LazyMatrix, lazily.}
\usage{
\S4method{t}{LazyMatrix}(x)
}
\description{
Transpose a LazyMatrix, lazily.
}
