% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\name{tcrossprod,LazyMatrix,missing-method}
\alias{tcrossprod,LazyMatrix,missing-method}
\title{tcrossprod a LazyMatrix, lazily.}
\usage{
\S4method{tcrossprod}{LazyMatrix,missing}(x)
}
\arguments{
\item{x}{LazyMatrix

This function is LAZY: it defers calculations to later.}
}
\description{
tcrossprod a LazyMatrix, lazily.
}
