% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\name{colSums,LazyMatrix-method}
\alias{colSums,LazyMatrix-method}
\title{Get sums of columns for LazyMatrix.}
\usage{
\S4method{colSums}{LazyMatrix}(x)
}
\arguments{
\item{x}{LazyMatrix

This function is EAGER: it triggers immediate calculations.}
}
\description{
Get sums of columns for LazyMatrix.
}
