% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval.R, R/MatrixLazyEval_core.R
\name{EvaluateLazyMatrix}
\alias{EvaluateLazyMatrix}
\alias{EvaluateLazyMatrix}
\title{Compute a matrix product.}
\usage{
EvaluateLazyMatrix(M, LEFT = Matrix::Diagonal(1, n = nrow(M)),
  RIGHT = Matrix::Diagonal(1, n = ncol(M)))

EvaluateLazyMatrix(M, LEFT = Matrix::Diagonal(1, n = nrow(M)),
  RIGHT = Matrix::Diagonal(1, n = ncol(M)))
}
\arguments{
\item{M}{object of class MatrixLazyEval.}

\item{LEFT}{x in xMy.}

\item{RIGHT}{y in xMy.}

\item{M}{object of class LazyMatrix.}

\item{RIGHT}{y in xMy.}

\item{LEFT}{x in xMy.}
}
\description{
Compute a matrix product.

Compute a matrix product.
}
