% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_core.R
\name{ExtractElementsLazyMatrix}
\alias{ExtractElementsLazyMatrix}
\title{Evaluate elements of a LazyMatrix.}
\usage{
ExtractElementsLazyMatrix(M, i, j, drop = T)
}
\arguments{
\item{M}{LazyMatrix}

\item{i, j}{Integer vectors.}

\item{drop}{If TRUE, return something 1d. Otherwise, return something 2d.}
}
\description{
Evaluate elements of a LazyMatrix.
}
