% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{methods}
\name{\%*\%,ANY,LazyMatrix-method}
\alias{\%*\%,ANY,LazyMatrix-method}
\title{Multiply a LazyMatrix by a regular hard-working matrix.}
\usage{
\S4method{\%*\%}{ANY,LazyMatrix}(x, y)
}
\arguments{
\item{x}{Anything with a matrix multiplication method that accepts objects in the components slot of y.}

\item{y}{LazyMatrix}
}
\description{
Multiply a LazyMatrix by a regular hard-working matrix.
}
