% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{methods}
\name{\%*\%,LazyMatrix,ANY-method}
\alias{\%*\%,LazyMatrix,ANY-method}
\title{Multiply a LazyMatrix by a regular hard-working matrix.}
\usage{
\S4method{\%*\%}{LazyMatrix,ANY}(x, y)
}
\arguments{
\item{x}{LazyMatrix}

\item{y}{Anything with a matrix multiplication method that accepts objects in the components slot of x.}
}
\description{
Multiply a LazyMatrix by a regular hard-working matrix.
}
