% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{methods}
\name{\%*\%,LazyMatrix,ANY-method}
\alias{\%*\%,LazyMatrix,ANY-method}
\title{Multiply a LazyMatrix by a regular hard-working matrix.}
\usage{
\S4method{\%*\%}{LazyMatrix,ANY}(x, y)
}
\description{
Multiply a LazyMatrix by a regular hard-working matrix.
}
