% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\name{rowSums,LazyMatrix-method}
\alias{rowSums,LazyMatrix-method}
\title{Get sums of rows for LazyMatrix.}
\usage{
\S4method{rowSums}{LazyMatrix}(x)
}
\arguments{
\item{x}{LazyMatrix

This function is EAGER: it triggers immediate calculations.}
}
\description{
Get sums of rows for LazyMatrix.
}
