% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\docType{methods}
\name{[,LazyMatrix,ANY,ANY,ANY-method}
\alias{[,LazyMatrix,ANY,ANY,ANY-method}
\title{Abdicate responsibility for weird multiplication requests.}
\usage{
\S4method{[}{LazyMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
Abdicate responsibility for weird multiplication requests.
}
