% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_methods.R
\name{t,LazyMatrix-method}
\alias{t,LazyMatrix-method}
\title{Transpose a LazyMatrix, lazily.}
\usage{
\S4method{t}{LazyMatrix}(x)
}
\arguments{
\item{x}{LazyMatrix

This function is LAZY: it defers calculations to later.}
}
\description{
Transpose a LazyMatrix, lazily.
}
