% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixLazyEval_core.R
\name{TestLazyMatrix}
\alias{TestLazyMatrix}
\title{Check matrix by computing xM, My, and xMy with random dense x, y.}
\usage{
TestLazyMatrix(M, seed = 0)
}
\arguments{
\item{M}{Lazymatrix}

\item{seed}{Seed for RNG.}
}
\description{
Check matrix by computing xM, My, and xMy with random dense x, y.
}
\details{
This does no more than make sure a quadratic form can be computed and that it yields a scalar as expected.
}
